feat: support using maru container in e2e tests (#116)
* feat: support using maru container in e2e tests

* feat: added log for useMaruContainer

* fix: spotless

* feat: add maru smoke test

* feat: revise shell script and job name

* feat: add workflow output and add debug log

* feat: temp run the build testing and smoke test unconditionally

* fix: added back develop_tag as workflow input

* feat: add debug step

* feat: add genesis-maru.json.template and fix maru tag env

* feat: revise debug step

* feat: revise debug step

* feat: update maru config and genesis

* feat: update if for build-and-publish

* feat: revise workflow for smoke testing

* feat: revise workflow for smoke testing

* feat: remove comment out job

* feat: revise based on comments

* fix: spotless

+++ maru/docker/initialization/init.sh
@@ -1,16 +1,18 @@
 #!/bin/zsh
 echo "Checking if genesis files already exist."
 
-if [[ ! -f "initialization/genesis-besu.json" && ! -f "initialization/genesis-geth.json" && ! -f "initialization/genesis-nethermind.json" ]]; then
-    echo "Initialization of timestamp in genesis files for Besu, Geth, and Nethermind."
+if [[ ! -f "initialization/genesis-besu.json" && ! -f "initialization/genesis-geth.json" && ! -f "initialization/genesis-nethermind.json"  && ! -f "initialization/genesis-maru.json" ]]; then
+    echo "Initialization of timestamp in genesis files for Maru, Besu, Geth, and Nethermind."
     date
     cd initialization || exit
+    cp -T "genesis-maru.json.template" "genesis-maru.json"
     cp -T "genesis-besu.json.template" "genesis-besu.json"
     cp -T "genesis-geth.json.template" "genesis-geth.json"
     cp -T "genesis-nethermind.json.template" "genesis-nethermind.json"
 
     merge_timestamp=$(($(date +%s) + 60))
     echo "Timestamp: $merge_timestamp"
+    sed -i "s/%PRAGUE_TIME%/$merge_timestamp/g" genesis-maru.json
     sed -i "s/%SWITCH_TIME%/$merge_timestamp/g" genesis-besu.json
     sed -i "s/%SWITCH_TIME%/$merge_timestamp/g" genesis-geth.json
     merge_timestamp_hex=$(printf "0x%x" $merge_timestamp)
